multilevel-queue-scheduling
Write a program for multilevel queue scheduling algorithm. There must be three queues generated. There must be a specific range of priority associated with every queue. Now prompt the user to enter the number of processes along with their priority and burst time. Each process must occupy the respective queue with specific priority range according to its priority. Apply Round robin algorithm with quantum time 4 on queue with the highest priority range. Apply priority scheduling algorithm on the queue with a medium range of priority and First come first serve algorithm on the queue with the lowest range of priority. Each and every queue should get a quantum time of 10 seconds. Cpu will keep on shifting between queues after every 10 seconds i.e. to apply round-robin algorithm OF 10 seconds on the overall structure. Calculate Waiting time and turnaround time for every process. The input for a number of processes should be given by the user.